adding setting for en21x release version
